% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_6400.R
\name{read_6400}
\alias{read_6400}
\title{Reads files from the LI-6400}
\usage{
read_6400(filename, data_start = 27)
}
\arguments{
\item{filename}{path and name of your data.}

\item{data_start}{the start of your data(without headline)}
}
\value{
read_6400 imports a LI-6400 raw data file as a data frame
}
\description{
\code{read_6400} Reads LI-6400 raw data files, which are delimited by tabs.
}
