% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bat_6400.R
\name{read_bat_6400}
\alias{read_bat_6400}
\title{read all raw data in one file measured by LI-6400}
\usage{
read_bat_6400(file_dir, header_line = 17, data_start = 27)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
raw data.}

\item{header_line}{ther start of your data with header.}

\item{data_start}{the start of your data(without headline)}
}
\description{
help to read all raw data files with a command.
}
\examples{
\dontrun{
library(readphoto)
read_bat_6400s('./6400')
}

}
