% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bat_6800.R
\name{read_bat_6800}
\alias{read_bat_6800}
\title{read all raw data in one file measured by LI-6800}
\usage{
read_bat_6800(file_dir, data_start = 56)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
raw data.}

\item{data_start}{the start of your data(without headline)}
}
\description{
help to read all raw data files with a command.
}
\examples{
\dontrun{
library(readphoto)
read_bat_6800('./6800')
}

}
