% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_regex68.R
\name{read_regex68}
\alias{read_regex68}
\title{read all raw data in one file measured by LI-6800}
\usage{
read_regex68(file_dir)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
raw data.}
}
\description{
help to read all raw data files with a command.
}
\examples{
\dontrun{
library(readphoto)
read_regex('./6800')
}

}
