% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl_6800.R
\name{readxl_6800}
\alias{readxl_6800}
\title{Reads files from the LI-6800}
\usage{
readxl_6800(filename, data_start = 41)
}
\arguments{
\item{filename}{path and name of your data.}

\item{data_start}{the start of your data(without headline)}
}
\value{
read_6800 imports a LI-6800 excel data file as a data frame
}
\description{
\code{read_6800} Reads LI-6800 excel data files, which are delimited by tabs.
}
