% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl_bat_6800.R
\name{readxl_bat_6800}
\alias{readxl_bat_6800}
\title{read all excel data in one file measured by LI-6800}
\usage{
readxl_bat_6800(file_dir, data_start = 41)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
excel data.}

\item{data_start}{the start of your data(without headline)}
}
\description{
help to read all excel data files with a command.
}
\examples{
\dontrun{
library(readphoto)
read_bat_6800s('./6800')
}

}
