% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recomp_6400.R
\name{recomp_6400}
\alias{recomp_6400}
\title{recompute the raw data in one file measured by LI-6400}
\usage{
recomp_6400(file_dir, header_line = 17, data_start = 27, S = 6, K = 0.5)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
raw data.}

\item{header_line}{ther start of your data with header.}

\item{data_start}{the start of your data(without headline)}

\item{S}{area of the leaf measured}

\item{K}{stomatal ratio}
}
\description{
help to read all raw data files with a command.
}
\examples{
\dontrun{
library(readphoto)
recomp_6400('./6400')
}

}
