% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recompxl_6800.R
\name{recompxl_6800}
\alias{recompxl_6800}
\title{recompute the excel data in one file measured by LI-6800}
\usage{
recompxl_6800(file_dir, data_start = 41, S = 6, K = 0.5)
}
\arguments{
\item{file_dir}{is the file directory only contains all the measured
excel data.}

\item{data_start}{the start of your data(without headline)}

\item{S}{area of the leaf measured}

\item{K}{stomatal ratio}
}
\description{
help to read all excel data files with a command.
}
\examples{
\dontrun{
library(readphoto)
recomp_6800('./6800')
}

}
