% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_read.R
\name{regex_read}
\alias{regex_read}
\title{Reads files from the LI-6800}
\usage{
regex_read(filename)
}
\arguments{
\item{filename}{path and name of your data.}
}
\value{
read_6800 imports a LI-6800 raw data file as a data frame
}
\description{
\code{read_6800} Reads LI-6800 raw data files, which are delimited by tabs.
}
